Move some RepoMan code around
