+++ Libs/libMacGitverCore/RepoMan/Frontend/Remote.cpp
@@ -19,13 +19,15 @@
 
 #include "libGitWrap/Result.hpp"
 
-#include "Repo.hpp"
-#include "Remote.hpp"
-#include "Events.hpp"
-#include "CollectionNode.hpp"
+#include "RepoMan/Events.hpp"
+
+#include "RepoMan/Repo.hpp"
+#include "RepoMan/Remote.hpp"
+#include "RepoMan/CollectionNode.hpp"
 
 #include "RepoMan/Private/Dumper.hpp"
-#include "RepoMan/Private/RemotePrivate.hpp"
+
+#include "RepoMan/Data/RemoteData.hpp"
 
 namespace RM
 {
