+++ Libs/libMacGitverCore/RepoMan/Frontend/Tag.cpp
@@ -17,11 +17,13 @@
  *
  */
 
-#include "Tag.hpp"
-#include "Events.hpp"
+#include "RepoMan/Events.hpp"
+
+#include "RepoMan/Tag.hpp"
 
 #include "RepoMan/Private/Dumper.hpp"
-#include "RepoMan/Private/TagPrivate.hpp"
+
+#include "RepoMan/Data/TagData.hpp"
 
 namespace RM
 {
